Adjustments to menu
