+++ src/frontend/Navigation/Stack/AppScreens.tsx
@@ -100,6 +100,7 @@ import {ObservationMetadata} from '../../screens/ObservationMetadata.tsx';
 import {ErrorBottomSheet} from '../../sharedComponents/ErrorBottomSheet.tsx';
 import {BackgroundMapErrorBottomSheet} from '../../screens/Settings/MapManagement/ErrorBottomSheet.tsx';
 import {MenuScreen} from '../../screens/MenuScreen';
+import {MenuHeader} from '../../sharedComponents/MenuHeader.tsx';
 
 export const TAB_BAR_HEIGHT = 70;
 
@@ -457,8 +458,9 @@ export const createDefaultScreenGroup = ({
         name="Menu"
         component={MenuScreen}
         options={{
+          headerShown: true,
           animation: 'slide_from_right',
-          headerShown: false,
+          header: () => <MenuHeader />,
         }}
       />
     </RootStack.Group>
