+++ src/frontend/sharedComponents/MenuList/MenuWrapper.tsx
@@ -1,0 +1,45 @@
+import React from 'react';
+import {
+  StyleSheet,
+  TouchableOpacity,
+  ViewStyle,
+  StyleProp,
+  TouchableOpacityProps,
+} from 'react-native';
+
+type MenuWrapperProps = {
+  onPress?: TouchableOpacityProps['onPress'];
+  children: React.ReactNode;
+  style?: StyleProp<ViewStyle>;
+  testID?: string;
+  accessibilityLabel?: string;
+};
+
+export function MenuWrapper({
+  onPress,
+  children,
+  style,
+  testID,
+  accessibilityLabel,
+}: MenuWrapperProps) {
+  return (
+    <TouchableOpacity
+      testID={testID}
+      accessibilityLabel={accessibilityLabel}
+      onPress={onPress}
+      activeOpacity={0.7}
+      style={[styles.container, style]}>
+      {children}
+    </TouchableOpacity>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    minHeight: 48,
+    paddingHorizontal: 15,
+    paddingVertical: 10,
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+});
